SpinBox class: Move to src/gui/mrview
This isn't used in shview, so might as well live here instead.
