+++ src/gui/mrview/tool/overlay.h
@@ -27,7 +27,7 @@
 #include "gui/mrview/tool/base.h"
 #include "gui/mrview/adjust_button.h"
 #include "gui/mrview/colourmap_button.h"
-#include "gui/spin_box.h"
+#include "gui/mrview/spin_box.h"
 
 namespace MR
 {
